volume issue fix

+++ Aeroclub.Cargo/Aeroclub.Cargo.Application/Models/RequestModels/PackageItemRMs/PackageDetailsUpdateRM.cs
@@ -10,7 +10,7 @@ namespace Aeroclub.Cargo.Application.Models.RequestModels.PackageItemRMs
     {
         public int width { get; set; }
         public int height { get; set; }
-        public int weight { get; set; }
+        public double weight { get; set; }
         public int length { get; set; }
         public string refNo {  get; set; }
